Handle the outdir option consistently
The outdir option was being expanded to the full path for the purposes of checking output files, but the relative path was being passed into the asn1 compiler. This created some unexpected results when the plugin was being used to compile a dependency. The output files were being placed in a directory at the top level of the project when they should be in a directory relative to the dependency.

Also, use 'filelib:ensure_dir' to create parent directories, if necessary.